interim groupnotefunction (based on roles)
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@748 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ classes/class.NoteFolderObject.php
@@ -38,26 +38,40 @@ class NoteFolderObject
 	* add note to notefolder 
 	* 
 	* @param	string  note_id
+	* @param	string  group_id = obj_id of group
 	* @access	public
 	*/
-	function addNote($note_id)
+	function addNote($note_id, $group="")
 	{
-/*
-		global $rbacadmin;
 
-		$roles = $rbacadmin->getRoleData(247); 
-		print_r($roles);
-*/
+		global $rbacreview;
+		//getgroupmembers
+		$grp_members = $rbacreview->assignedUsers($group);
 	
-		$myTree = new tree(0, 0, 0, $this->m_usr_id); 
+		if(!empty($group))
+		{
+			foreach($grp_members as $member)
+			{	
+				$myTree = new tree(0, 0, 0, $member); 
 
+				//get parent_id of usersettingfolder...	
+				$rootid =  $myTree->getRootID($member);
+	
+				$node_data = $myTree->getNodeDataByType("notf");
+
+				$myTree->insertNode($note_id, $node_data[0]["obj_id"], $rootid["child"]);
+			
+			}
+		}
+		//insert the note in ones own notefolder
+		$myTree = new tree(0, 0, 0, $this->m_usr_id); 
+	
 		//get parent_id of usersettingfolder...	
 		$rootid =  $myTree->getRootID($this->m_usr_id);
-
+	
 		$node_data = $myTree->getNodeDataByType("notf");
 
 		$myTree->insertNode($note_id, $node_data[0]["obj_id"], $rootid["child"]);
-
 	} 
 
 
@@ -69,20 +83,26 @@ class NoteFolderObject
 	*/
 	function deleteNotes($notes)
 	{
+		global $rbacsystem;
 		$myTree = new tree($this->m_notefId[0]["obj_id"], 0, 0, $this->m_usr_id); 	
 
 		foreach($notes as $note)
 		{
-			//just interim solution ->todo:object::deleteObject, so that object gets deleted in object_data
+			//delete note in note_data, only owner can delete notes 
+			$note_obj = getObject($note);
+			if($note_obj["owner"] == $this->m_usr_id)
+			{
+				$query = "DELETE FROM note_data WHERE note_id ='".$note."'";
+				$res = $this->ilias->db->query($query);
+				//TODO: delete entry in object_data and all dependencies, references 
+				//getgroupmembers and delete the note in members notefolder
+			}
 
 			//get note_data of note folder
 			$node_data1 = $myTree->getNodeDataByType("notf");		
 			$note_data2 = $myTree->getNodeData($note, $node_data1[0]["obj_id"]);	
 			$myTree->deleteTree($note_data2);				
-		
-			//delete note in note_data
-			$query = "DELETE FROM note_data WHERE note_id ='".$note."'";
-			$res = $this->ilias->db->query($query);
+
 		}
 	}
 
@@ -105,7 +125,7 @@ class NoteFolderObject
 		{
 			foreach($nodes as $node_data)
 			{
-				$note_data = NoteObject::viewObject($node_data["child"]);
+				$note_data = getObject($node_data["child"]);
 				array_push($notes, $note_data);
 			}
 
@@ -113,11 +133,21 @@ class NoteFolderObject
 		else
 		{	
 			$node_data["child"] = $note_id;
-			$notes = NoteObject::viewObject($node_data["child"]);		
+			$notes = getObject($node_data["child"]);
 		}
 		return $notes;					
 	}
 
+	function viewNote($note_id)
+	{
+		$note = array();
+		$myTree = new tree($this->m_notefId[0]["obj_id"], 0, 0, $this->m_usr_id); 	
+	
+		$nodes = $myTree->getNodeDataByType("note");
+		$node_data["child"] = $note_id;
+		$note = NoteObject::viewObject($node_data["child"]);		
+		return $note;	
+	}
 
 }
